More drafts the oracle can judge: labels, indexed loads, mixed registers
Three more reasons a draft never reached tools/match_function.py, each
found by recording the compiler diagnostic for drafts that fail to build
rather than just counting them (tools/recarve_drafts.py --error-log).

1. enumerate_functions synthesises `func_ADDRESS` for a range splat left
   without a glabel, which happens where a function follows a `jr $ra`
   tail splat did not treat as a boundary. m2c looks functions up by
   label, so it answered "Function not found" for 76 spans and they got no
   draft at all. m2c_assembly_text now injects the label into m2c's
   private copy; asm/ is untouched. func_8005080C, 2,104 bytes, is one of
   them.

2. `*(D + i)`, m2c's spelling of a load through a label with a computed
   displacement, does not compile against the sanitizer's `extern s32 D;`
   default. It was the single most common failure, 43 of 70 sampled. The
   fix is the array form `extern s32 D[];`, not the pointer form: an array
   decays to the label's own address, which is what retail's `%lo(D)(reg)`
   addressing does, whereas `extern s32 *D;` would load a pointer out of
   that address and offset from there. That compiles and is wrong, which
   is worse than not compiling.

3. The same problem one scope down, for locals and parameters holding an
   address. Where m2c uses one register as a pointer on one line and an
   integer on the next, the declaration is not the place to guess, so the
   cast goes at the dereference site and both readings survive as m2c
   wrote them.

Also widened recarve_drafts to the spans that carry only a word export and
a note saying m2c produced nothing. Those claim nothing, so replacing the
body cannot lose work.

Between them, 110 more drafts now compile. The honest result is that this
buys much less than the first round did: the sweep found 3 new matches, and
the mismatch population grew from 827 to 892. That is the expected shape.
Compilability was a floor, not a ceiling, and the remaining population is
"compiles but is not retail", which needs per-function work rather than
another systemic fix.

Coverage 247,840 -> 248,016 / 806,272 bytes (30.74% -> 30.76%).
verify_registry green at 2,647/2,647; 1,601 source-sensitive tests pass.

+++ src/overlays/main_0012/801598e0.c
@@ -59,7 +59,7 @@ void func_801598E0(void *arg0) {
         break;
     case 0x53:
     case 0x73:
-        M2C_FIELD(arg0, u8 *, 0xAE) = (s16) (M2C_FIELD(arg0, u8 *, 0xAE) | 0x8000);
+        M2C_FIELD(arg0, s16 *, 0xAE) = (s16) (M2C_FIELD(arg0, u8 *, 0xAE) | 0x8000);
         break;
     }
     func_80159968();
